+++ apps/backend/src/modules/project-groups/libs/types/types.ts
@@ -3,4 +3,6 @@ export {
 	type ProjectGroupCreateResponseDto,
 	type ProjectGroupGetAllRequestDto,
 	type ProjectGroupGetAllResponseDto,
+	type ProjectGroupPatchRequestDto,
+	type ProjectGroupPatchResponseDto,
 } from "@git-fit/shared";