+++ apps/backend/src/modules/users/user.controller.ts
@@ -1,4 +1,5 @@
-import { APIPath } from "~/libs/enums/enums.js";
+import { APIPath, PermissionKey } from "~/libs/enums/enums.js";
+import { checkUserPermissions } from "~/libs/hooks/check-user-permissions.hook.js";
 import {
 	type APIHandlerOptions,
 	type APIHandlerResponse,
@@ -51,6 +52,7 @@ class UserController extends BaseController {
 				),
 			method: "GET",
 			path: UsersApiPath.ROOT,
+			preHandlers: [checkUserPermissions([PermissionKey.MANAGE_USER_ACCESS])],
 		});
 
 		this.addRoute({
@@ -63,6 +65,7 @@ class UserController extends BaseController {
 				),
 			method: "PATCH",
 			path: UsersApiPath.$ID,
+			preHandlers: [checkUserPermissions([PermissionKey.MANAGE_USER_ACCESS])],
 			validation: {
 				body: userPatchValidationSchema,
 			},
@@ -77,6 +80,7 @@ class UserController extends BaseController {
 				),
 			method: "DELETE",
 			path: UsersApiPath.$ID,
+			preHandlers: [checkUserPermissions([PermissionKey.MANAGE_USER_ACCESS])],
 		});
 	}