+++ apps/frontend/src/pages/project-access-management/libs/components/components.ts
@@ -1,4 +1,5 @@
 export { ProjectGroupCreateForm } from "./project-group-create-form/project-group-create-form.js";
 export { ProjectGroupMenu } from "./project-group-menu/project-group-menu.js";
+export { ProjectGroupUpdateForm } from "./project-group-update-form/project-group-update-form.js";
 export { ProjectGroupsTable } from "./project-groups-table/project-groups-table.js";
 export { UsersTable } from "./users-table/users-table.js";