+++ apps/frontend/src/pages/project/libs/components/contributors-list/contributors-list.tsx
@@ -1,4 +1,6 @@
 import { Loader } from "~/libs/components/components.js";
+import { EMPTY_LENGTH } from "~/libs/constants/constants.js";
+import { useCallback } from "~/libs/hooks/hooks.js";
 import { type ContributorGetAllItemResponseDto } from "~/pages/project/libs/types/types.js";
 
 import { ContributorCard } from "../components.js";
@@ -7,25 +9,46 @@ import styles from "./styles.module.css";
 type Properties = {
 	contributors: ContributorGetAllItemResponseDto[];
 	isLoading: boolean;
+	onEditContributor: (contributorId: number) => void;
 };
 
 const ContributorsList = ({
 	contributors,
 	isLoading,
+	onEditContributor,
 }: Properties): JSX.Element => {
+	const hasContributors = contributors.length > EMPTY_LENGTH;
+	const isListShown = !isLoading && hasContributors;
+	const isEmptyPlaceholderShown = !isLoading && !hasContributors;
+
+	const handleEditContributor = useCallback(
+		(contributorId: number) => {
+			onEditContributor(contributorId);
+		},
+		[onEditContributor],
+	);
+
 	return (
 		<div className={styles["container"]}>
 			<h2 className={styles["title"]}>Contributors</h2>
-			{isLoading ? (
-				<Loader />
-			) : (
+			{isLoading && <Loader />}
+			{isListShown && (
 				<ul className={styles["list"]}>
 					{contributors.map((contributor) => (
 						<li key={contributor.id}>
-							<ContributorCard contributor={contributor} />
+							<ContributorCard
+								contributor={contributor}
+								onEdit={handleEditContributor}
+							/>
 						</li>
 					))}
 				</ul>
+			)}
+			{isEmptyPlaceholderShown && (
+				<p className={styles["empty-placeholder"]}>
+					There are no contributors at the moment. Please set up analytics if
+					you haven&#39;t already.
+				</p>
 			)}
 		</div>
 	);