+++ apps/frontend/src/pages/project/libs/components/setup-analytics-modal/setup-analytics-modal.tsx
@@ -11,9 +11,11 @@ import {
 	useAppSelector,
 	useCallback,
 	useEffect,
+	useMemo,
 } from "~/libs/hooks/hooks.js";
 import { actions as projectApiKeyActions } from "~/modules/project-api-keys/project-api-keys.js";
 import { type ProjectGetByIdResponseDto } from "~/modules/projects/projects.js";
+import { actions as scriptActions } from "~/modules/scripts/scripts.js";
 
 import styles from "./styles.module.css";
 
@@ -29,18 +31,38 @@ const SetupAnalyticsModal = ({
 	project,
 }: Properties): JSX.Element => {
 	const dispatch = useAppDispatch();
-	const { control, errors, handleSubmit, handleValueSet } = useAppForm({
-		defaultValues: {
-			apiKey: project.apiKey ?? "",
-			projectId: project.id,
-		},
-	});
+
 	const { dataStatus } = useAppSelector(({ projectApiKeys }) => projectApiKeys);
+	const { authenticatedUser } = useAppSelector(({ auth }) => auth);
 
 	const hasProjectApiKey = project.apiKey !== null;
+	const hasAuthenticatedUser = authenticatedUser !== null;
+
 	const isGenerateButtonDisabled = dataStatus === DataStatus.PENDING;
 	const isCopyButtonDisabled =
 		!hasProjectApiKey || dataStatus === DataStatus.PENDING;
+
+	const pm2StartupScript = "pm2 startup";
+
+	const analyticsScript = useMemo<string>(() => {
+		if (!hasProjectApiKey || !hasAuthenticatedUser) {
+			return "";
+		}
+
+		const apiKey = project.apiKey as string;
+		const userId = String(authenticatedUser.id);
+
+		return `npx @git-fit/analytics@latest track ${apiKey} ${userId} <project-path>`;
+	}, [hasProjectApiKey, hasAuthenticatedUser, project, authenticatedUser]);
+
+	const { control, errors, handleSubmit, handleValueSet } = useAppForm({
+		defaultValues: {
+			analyticsScript,
+			apiKey: project.apiKey ?? "",
+			pm2StartupScript,
+			projectId: project.id,
+		},
+	});
 
 	const handleGenerateSubmit = useCallback(
 		(event_: React.BaseSyntheticEvent): void => {
@@ -51,19 +73,51 @@ const SetupAnalyticsModal = ({
 		[handleSubmit, dispatch],
 	);
 
-	const handleCopyToClipboardClick = useCallback(() => {
-		void dispatch(
-			projectApiKeyActions.copyToClipboard(project.apiKey as string),
-		);
-	}, [dispatch, project.apiKey]);
+	const handleCopyApiKeyToClipboard = useCallback(
+		(input: string) => {
+			void dispatch(projectApiKeyActions.copyToClipboard(input));
+		},
+		[dispatch],
+	);
+
+	const handleCopyScriptToClipboard = useCallback(
+		(input: string) => {
+			void dispatch(scriptActions.copyToClipboard(input));
+		},
+		[dispatch],
+	);
+
+	const handleCopyAPIKeyClick = useCallback(() => {
+		handleCopyApiKeyToClipboard(project.apiKey as string);
+	}, [handleCopyApiKeyToClipboard, project]);
+
+	const handleCopyAnalyticsScriptClick = useCallback(() => {
+		handleCopyScriptToClipboard(analyticsScript);
+	}, [handleCopyScriptToClipboard, analyticsScript]);
+
+	const handleCopyStartupScriptClick = useCallback(() => {
+		handleCopyScriptToClipboard(pm2StartupScript);
+	}, [handleCopyScriptToClipboard, pm2StartupScript]);
 
 	useEffect(() => {
 		handleValueSet("apiKey", project.apiKey ?? "");
 	}, [handleValueSet, project.apiKey]);
 
+	useEffect(() => {
+		handleValueSet("analyticsScript", analyticsScript);
+	}, [handleValueSet, analyticsScript]);
+
 	return (
 		<Modal isOpened={isOpened} onClose={onClose} title="Setup Analytics">
 			<div className={styles["content"]}>
+				<div>
+					<span className={styles["subtitle"]}>Overview</span>
+					<p className={styles["text"]}>
+						This script operates continuously in the background, collecting
+						statistics and updating analytics data every 3 hours.
+					</p>
+				</div>
+
 				<form
 					className={styles["api-key-container"]}
 					onSubmit={handleGenerateSubmit}
@@ -80,7 +134,7 @@ const SetupAnalyticsModal = ({
 								iconName="clipboard"
 								isDisabled={isCopyButtonDisabled}
 								label="Copy API key"
-								onClick={handleCopyToClipboardClick}
+								onClick={handleCopyAPIKeyClick}
 							/>
 						}
 					/>
@@ -92,6 +146,112 @@ const SetupAnalyticsModal = ({
 						/>
 					</div>
 				</form>
+
+				<div>
+					<span className={styles["subtitle"]}>Prerequisites</span>
+					<ul className={styles["text"]}>
+						<li>
+							<span className={styles["list-item-title"]}>Git</span>: Ensure Git
+							is installed for repository management.
+						</li>
+						<li>
+							<span className={styles["list-item-title"]}>Node.js 20</span>: The
+							script requires Node.js 20 to be installed on your machine.
+						</li>
+						<li>
+							<span className={styles["list-item-title"]}>PM2 5.4</span>: The
+							script requires PM2 5.4 to be installed on your machine.
+						</li>
+						<li>
+							<span className={styles["list-item-title"]}>
+								Unix-based system
+							</span>
+							: The script requires a Unix-based operating system (e.g., Linux
+							or macOS) to run properly.
+						</li>
+					</ul>
+				</div>
+
+				<div>
+					<span className={styles["subtitle"]}>Installation Steps</span>
+					<ol className={styles["text"]}>
+						<li className={styles["list-item"]}>
+							<span className={styles["list-item-title"]}>
+								Execute the configuration script.
+							</span>
+							<p className={styles["list-item-text"]}>
+								Open your terminal or console, copy the following script, paste
+								and run it.
+							</p>
+							<Input
+								control={control}
+								errors={errors}
+								isLabelHidden
+								isReadOnly
+								label="pm2StartupScript"
+								name="pm2StartupScript"
+								placeholder="Need to generate API key"
+								rightIcon={
+									<IconButton
+										iconName="clipboard"
+										isDisabled={isCopyButtonDisabled}
+										label="Copy script"
+										onClick={handleCopyStartupScriptClick}
+									/>
+								}
+							/>
+							<span className={styles["list-item-text"]}>
+								Then, modify the command output from this step with your
+								system&apos;s values and execute it in terminal or console.
+							</span>
+						</li>
+						<li className={styles["list-item"]}>
+							<span className={styles["list-item-title"]}>
+								Clone your project repository.
+							</span>
+							<p className={styles["list-item-text"]}>
+								Use Git to clone your project repository to your local machine.
+							</p>
+						</li>
+
+						<li className={styles["list-item"]}>
+							<span className={styles["list-item-title"]}>
+								Prepare the script.
+							</span>
+							<p className={styles["list-item-text"]}>
+								Copy the command below and replace &lt;project-path&gt;
+								placeholder with your local repository&apos;s path:
+							</p>
+							<Input
+								control={control}
+								errors={errors}
+								isLabelHidden
+								isReadOnly
+								label="Analytics script"
+								name="analyticsScript"
+								placeholder="Need to generate API key"
+								rightIcon={
+									<IconButton
+										iconName="clipboard"
+										isDisabled={isCopyButtonDisabled}
+										label="Copy script"
+										onClick={handleCopyAnalyticsScriptClick}
+									/>
+								}
+							/>
+						</li>
+
+						<li className={styles["list-item"]}>
+							<span className={styles["list-item-title"]}>
+								Execute the script.
+							</span>
+							<p className={styles["list-item-text"]}>
+								Open your terminal or console, paste and run the modified
+								script. Script will start and be saved to restart on reboot.
+							</p>
+						</li>
+					</ol>
+				</div>
 			</div>
 		</Modal>
 	);