+++ packages/shared/src/modules/contributors/libs/types/contributor-get-all-item-response-dto.type.ts
@@ -4,6 +4,7 @@ type ContributorGetAllItemResponseDto = {
 		id: number;
 	}[];
 	id: number;
+	lastActivityDate: null | string;
 	name: string;
 	projects: {
 		id: number;