+++ packages/shared/src/modules/contributors/libs/types/contributor-patch-request-dto.type.ts
@@ -1,0 +1,5 @@
+type ContributorPatchRequestDto = {
+	name: string;
+};
+
+export { type ContributorPatchRequestDto };