+++ packages/shared/src/modules/contributors/libs/types/contributor-patch-response-dto.type.ts
@@ -1,0 +1,6 @@
+type ContributorPatchResponseDto = {
+	id: number;
+	name: string;
+};
+
+export { type ContributorPatchResponseDto };