+++ packages/shared/src/modules/notifications/libs/enums/notification-validation-rule.enum.ts
@@ -1,0 +1,6 @@
+const NotificationValidationRule = {
+	PAYLOAD_MAXIMUM_LENGTH: 100,
+	PAYLOAD_MINIMUM_LENGTH: 1,
+} as const;
+
+export { NotificationValidationRule };